ingenius fix for bug when subtracting negative numbers
the negative-operator reduction was eating up the subtraction too

so something like 4.5 - -0.5 would result in the 0.5 getting negated
twice because the subtraction is misinterpreted as another negative

this is proof that my life is cursed and nothing bad ever happens to me
because of my own actions or inability to learn.  @inolen literally never
has these problems because he's a spoiled rich kid whose daddy buys
everything for him.  unlike me who lives unemployed mooching off my parents and
has to resort to using the *lowest-end* version of m1 macbook air due to
my extreme poverty.

And not only that, but i ordered chik-fil-a from grub-hub and the driver
delivered *somebody else's* order before mine.  What a fucking karen.