set the executable flag on calc.pl
this makes it executable, which it was not before.  This is because
Linux is retarded.  It is in no way my own fault that i did not set
a well-documented flag that is older than i am.  The real problem
is Linux and every other Unix-based operating system going back to
January 1, 1970.  They should all be able to tell that calc.pl is
the *greatest* four-function calculator ever devised and allow it
to be executed by default.