brilliant simplification of implicit multiplication
the old implementation worked for the cases of 'z(x + y)
and (w + x)(y + z) but it didn't work for the case of '(x + y)z'.

at first i was going to add in a special case for this, but being the
undiscovered programming prodigy that i am, i found that this could also be
solved by treating all instances where two products are directly adjacent to
one another as an implicit multiplication.  This would even be able to cover
cases such as xy == x*y, except there's no way to input that because there
are no variable substitutions in calc-tutorial.